+++ include/boost/compute/algorithm/copy.hpp
@@ -14,7 +14,7 @@
 #include <algorithm>
 #include <iterator>
 
-#include <boost/utility/enable_if.hpp>
+#include <boost/core/enable_if.hpp>
 
 #include <boost/mpl/and.hpp>
 #include <boost/mpl/not.hpp>
