+++ include/boost/compute/algorithm/sort_by_key.hpp
@@ -13,8 +13,8 @@
 
 #include <iterator>
 
+#include <boost/core/enable_if.hpp>
 #include <boost/static_assert.hpp>
-#include <boost/utility/enable_if.hpp>
 
 #include <boost/compute/system.hpp>
 #include <boost/compute/command_queue.hpp>
