+++ include/boost/compute/detail/is_contiguous_iterator.hpp
@@ -15,8 +15,8 @@
 #include <valarray>
 
 #include <boost/config.hpp>
+#include <boost/core/enable_if.hpp>
 #include <boost/type_traits.hpp>
-#include <boost/utility/enable_if.hpp>
 
 namespace boost {
 namespace compute {
