+++ include/boost/compute/iterator/buffer_iterator.hpp
@@ -15,9 +15,9 @@
 #include <iterator>
 
 #include <boost/config.hpp>
+#include <boost/core/enable_if.hpp>
 #include <boost/type_traits.hpp>
 #include <boost/static_assert.hpp>
-#include <boost/utility/enable_if.hpp>
 #include <boost/iterator/iterator_facade.hpp>
 
 #include <boost/compute/buffer.hpp>
